fix: moved classes related to unique gifts to appropriate packages
