+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/gifts/UniqueGiftBackdrop.kt
@@ -1,5 +1,8 @@
-package dev.inmo.tgbotapi.types
+package dev.inmo.tgbotapi.types.gifts
 
+import dev.inmo.tgbotapi.types.colorsField
+import dev.inmo.tgbotapi.types.nameField
+import dev.inmo.tgbotapi.types.rarityPerMilleField
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
